Extract printComment
